add remark-obsidian (for fun), ignore obsidian db
for now, i've ignored the obsidian db files. it may be useful to have them down the road. but let's start without.

also, i installed this just for fun. really don't know how useful it will be.

i've created a fork if some of these features need to be edited (like slug with removes slashes that i'll need for mi site structure)